Update addTask() and deleteTask()
Added functionality to insert a new task and display it in the listview. Also I can now select an item from the listView and hit the button delete to delete it. Still need to implement a "modify" option, and there's a bug: when deleting all the elements and add a new task, in the listview appears the new task and the last task inserted befor the deletion